+++ app/support/DbAdapter/index.d.ts
@@ -82,6 +82,7 @@ export class DbAdapter {
 
   // Users
   getUserById(id: UUID): Promise<User | null>;
+  getUserByIntId(intId: number): Promise<User | null>;
   getUsersByIds(ids: UUID[]): Promise<User[]>;
   getUserByUsername(username: string): Promise<User | null>;
   getUserIdsWhoBannedUser(id: UUID): Promise<UUID[]>;
@@ -187,9 +188,17 @@ export class DbAdapter {
     commentId?: Nullable<UUID>,
     postAuthorIntId?: Nullable<number>,
     targetPostId?: Nullable<UUID>,
-    targetCommntId?: Nullable<UUID>,
+    targetCommentId?: Nullable<UUID>,
   ): Promise<EventRecord>;
   getEventById(eventId: UUID): Promise<Nullable<EventRecord>>;
+  getUserEvents(
+    userIntId: number,
+    eventTypes?: string[],
+    limit?: number,
+    offset?: number,
+    startDate?: Date,
+    endDate?: Date,
+  ): Promise<EventRecord[]>;
 
   getUnreadDirectsNumber(userId: UUID): Promise<number>;
   getUnreadEventsNumber(userId: UUID): Promise<number>;